Use darker background color
Use the darker end of the background gradient for the Emacs site's
.header element.

For some reason, a linear-gradient isn't working correctly on Reddit:
instead of covering the whole page, it repeats 3 times down the page.
So it's necessary to use a solid color instead.  So we should use the
darker color, which gives a closer appearance to the Emacs site.